+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [2.0.6]
 
+### Features
+- Added an optional argument in `gameplay` tag to choose which gamemode you want the player to be `gameplay gamemode:survival`
+
 ### Fixes
 - Better retro-compatibility for animations
 - Can't access options screen from main screen in 1.21.1